[EMCAL-402] Cache ALTRO mapping in the TRU raw analyzer
The ALTRO raw stream is constructed for every event and
therefore always reads the same ALTRO mapping from a text
file from disk. As the altro mapping doesn't change it is
can be cached in memory and passed to the raw analyzer
from cache, which is a significantly faster operation than
reading it from disk.