[EMCAL-33]  Fix size of the input digit array
Input digit array must cover in the worst case the full
EMCAL + DCAL occupancy - not sufficient to adapt it to
one tracking sector.

In addition boundary checks for internal maps in case
of failures - should never happen.

For the digits monitor it has been observed in rare
cases that the gaintype was set incorrectly - hard fix
to 1 untill understood.